+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 
 
+## Unreleased
+
+### Added
+
+- Test error handler
+
+
 ## 0.1.0 - 2017-04-20
 
 ### Added